# apps/consumer/src/openapi.json
{"openapi":"3.0.0","paths":{"/api/contracts":{"get":{"operationId":"ContractsController_list","parameters":[{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}},{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Contracts"]},"post":{"operationId":"ContractsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContract"}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["Contracts"]}},"/api/contracts/{id}":{"put":{"operationId":"ContractsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateContract"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Contracts"]}},"/api/payments":{"get":{"operationId":"PaymentsController_list","parameters":[{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"]},"post":{"operationId":"PaymentsController_start","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartPayment"}}}},"responses":{"201":{"description":""}},"tags":["Payments"]}},"/api/payments/{id}/status":{"patch":{"operationId":"PaymentsController_patch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePaymentStatus"}}}},"responses":{"200":{"description":""}},"tags":["Payments"]}},"/api/documents":{"get":{"operationId":"DocumentsController_list","parameters":[{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentListItem"}}}}}},"tags":["Documents"]},"post":{"operationId":"DocumentsController_upload","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadDocument"}}}},"responses":{"201":{"description":""}},"tags":["Documents"]}},"/api/documents/presigned":{"post":{"operationId":"DocumentsController_presigned","parameters":[],"responses":{"201":{"description":""}},"tags":["Documents"]}},"/api/dashboard":{"get":{"operationId":"DashboardController_get","parameters":[{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Dashboard"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"tags":["Auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_whoami","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Auth"]}},"/api/auth/echo":{"get":{"operationId":"AuthController_echo","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/admins/users":{"get":{"operationId":"AdminsController_listUsers","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/users/{id}/role":{"patch":{"operationId":"AdminsController_setRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/contractors":{"get":{"operationId":"AdminsController_listContractors","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]},"post":{"operationId":"AdminsController_createContractor","parameters":[],"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/contractors/{id}":{"patch":{"operationId":"AdminsController_updateContractor","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]},"delete":{"operationId":"AdminsController_deleteContractor","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/contracts":{"get":{"operationId":"AdminsController_listContracts","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/contracts/{id}":{"patch":{"operationId":"AdminsController_updateContract","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]},"delete":{"operationId":"AdminsController_deleteContract","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/payments":{"get":{"operationId":"AdminsController_listPayments","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/payments/{id}":{"delete":{"operationId":"AdminsController_deletePayment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/documents":{"get":{"operationId":"AdminsController_listDocuments","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}},"/api/admins/documents/{id}":{"delete":{"operationId":"AdminsController_deleteDocument","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["Admins"]}}},"info":{"title":"Remoola API","description":"Client dashboard API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"jwt":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateContract":{"type":"object","properties":{"clientId":{"type":"string"},"contractorId":{"type":"string"},"rateCents":{"type":"number"},"rateUnit":{"type":"string"},"status":{"type":"string"}},"required":["clientId","contractorId","rateCents","rateUnit"]},"UpdateContract":{"type":"object","properties":{"rateCents":{"type":"number"},"rateUnit":{"type":"string"},"status":{"type":"string"}}},"StartPayment":{"type":"object","properties":{"contractId":{"type":"string"},"amountCents":{"type":"number"},"currency":{"type":"string"},"method":{"type":"string"}},"required":["contractId","amountCents"]},"UpdatePaymentStatus":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]},"DocumentListItem":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"size":{"type":"string"},"updated":{"type":"string"},"fileUrl":{"type":"string"}},"required":["id","name","type","size","updated"]},"UploadDocument":{"type":"object","properties":{"contractId":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"fileUrl":{"type":"string"},"sizeBytes":{"type":"number"}},"required":["contractId","name","type"]},"Login":{"type":"object","properties":{"email":{"type":"string","example":"remoola@example.com"},"password":{"type":"string","example":"remoola"}},"required":["email","password"]},"AuthResponse":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]}}}}